requirements.txt
# maintain alphabatical order
fastapi==0.103.2
loguru==0.7.2
numpy==1.26.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.1
pycocotools==2.0.7
uvicorn==0.23.2